refactor(core): Consolidate the node search engine into @n8n/ai-utilities (no-changelog)
The MCP tool surface and Instance AI each carried a near-identical node search
engine, ~500 lines apiece, with the same 8 methods and the same scoring keys.
Instance AI's copy existed because that package "intentionally does NOT depend on
n8n-workflow", which stopped being true a while ago: n8n-workflow is a declared
dependency and 13 non-test files already import from it.

Both now use one `NodeSearchEngine` in `@n8n/ai-utilities/node-catalog`. The node
and result types are widened to the union of what each side needed, so callers
pass whatever shape they already hold: a full `INodeTypeDescription`, a lean one,
or the leaner shape a host assembles.

Behaviour kept from Instance AI's copy, now shared:
- per-query result caching, with results copied on the way out
- `builderHint.message` preferred over `searchHint`
- n8n Connect metadata on results and the `<n8n_credits />` marker in output

Filtered `searchByName` calls skip the cache, since a predicate cannot be part of
a cache key.

Two fixes fall out of the merge. Instance AI's local AI connection type list had
drifted from `NodeConnectionTypes`: it invented `ai_vectorRetriever`, which no
node ever outputs, so searching it always returned nothing, and it omitted the
real `ai_reranker`, so rerankers were unreachable from the nodes tool. The
shared list now matches, and a `satisfies` clause fails the build if a value
that is not a real connection type is added again.

Test coverage is preserved: the Instance AI engine tests move across as
`search-engine.host-shape.test.ts`, covering the caching, n8n Connect and
`builderHint.message` paths that the existing suite did not.
