+++ packages/@n8n/instance-ai/src/tools/nodes.tool.ts
@@ -2,13 +2,17 @@
  * Consolidated nodes tool — list, search, describe, type-definition, suggested, explore-resources.
  */
 import { Tool } from '@n8n/agents';
-import { categoryList, suggestedNodesData } from '@n8n/ai-utilities/node-catalog';
+import {
+	AI_CONNECTION_TYPES,
+	NodeSearchEngine,
+	categoryList,
+	suggestedNodesData,
+	type SearchableNodeType,
+} from '@n8n/ai-utilities/node-catalog';
 import { z } from 'zod';
 
 import { sanitizeInputSchema } from '../agent/sanitize-mcp-schemas';
 import type { InstanceAiContext } from '../types';
-import { NodeSearchEngine } from './nodes/node-search-engine';
-import { AI_CONNECTION_TYPES, type SearchableNodeType } from './nodes/node-search-engine.types';
 import { pickPreferredChatModelNode } from './nodes/preferred-chat-model';
 import { buildCredentialMap } from './workflows/resolve-credentials';
